requirements.txt
python-telegram-bot==13.15
requests==2.31.0
pandas==2.1.1
numpy==1.27.5